requirements.txt
agno>=1.7.0
sympy>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0